GoldEater/requirements.txt
# GoldEater Dependencies

# AI Platform SDKs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# HTTP Client
requests>=2.31.0

# Database
supabase>=2.0.0

# Geo
h3>=4.0.0

# Utils
python-dotenv>=1.0.0
